use pre-defined icons in menu as well
