+++ tests/Twig/ExtensionsTest.php
@@ -50,7 +50,7 @@ class ExtensionsTest extends TestCase
 
     public function testGetFilters()
     {
-        $filters = ['duration', 'money', 'currency', 'country', 'icon', 'docu_link'];
+        $filters = ['duration', 'money', 'currency', 'country', 'docu_link'];
         $sut = $this->getSut($this->localeDe);
         $twigFilters = $sut->getFilters();
         $this->assertCount(count($filters), $twigFilters);
@@ -211,28 +211,6 @@ class ExtensionsTest extends TestCase
         $record->setDuration($seconds);
 
         return $record;
-    }
-
-    public function testIcon()
-    {
-        $icons = [
-            'user', 'customer', 'project', 'activity', 'admin', 'invoice', 'timesheet', 'dashboard', 'logout', 'trash',
-            'delete', 'repeat', 'edit', 'manual', 'help', 'start', 'start-small', 'stop', 'stop-small', 'filter',
-            'create', 'list', 'print', 'visibility', 'calendar', 'money', 'duration', 'download', 'copy', 'settings',
-            'export', 'pdf', 'csv', 'ods', 'xlsx', 'on', 'off', 'audit', 'home', 'shop', 'about', 'debug', 'profile-stats'
-        ];
-
-        // test pre-defined icons
-        $sut = $this->getSut($this->localeEn);
-        foreach ($icons as $icon) {
-            $result = $sut->icon($icon);
-            $this->assertNotEmpty($result, 'Problem with icon definition: ' . $icon);
-            $this->assertIsString($result);
-        }
-
-        // test fallback will be returned
-        $this->assertEquals('', $sut->icon('foo'));
-        $this->assertEquals('bar', $sut->icon('foo', 'bar'));
     }
 
     public function testDocuLink()
